+++ src/common.rs
@@ -24,3 +24,4 @@ pub trait ClipboardProvider: Send {
     /// Method to set the clipboard contents as a String
     fn set_contents(&mut self, _: String) -> Result<()>;
 }
+