+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 0.6.3
 
-### REmove
+### Added
 
 - Features `x11` and `wayland` for picking the linux backends