+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 # Unreleased
-
+- Updated `clipboard-win` to 4.x,x
+- Updated `x11-clipboard` to 0.6.x 
 - Minumun rust version was bumped to `1.56.0`
 
 ## 0.7.1